added yuansyuan's smooth feature rewind state is now NOT reseted after loadState

+++ src/win32/MainWnd.cpp
@@ -476,7 +476,7 @@ bool MainWnd::FileRun()
     return false;
   }
   systemSaveUpdateCounter = SYSTEM_SAVE_NOT_UPDATED;
-  theApp.cartridgeType = (int)type;
+  theApp.cartridgeType = type;
   if(type == IMAGE_GB) {
     genericflashcardEnable = theApp.winGenericflashcardEnable;
